rabbit mq to yii service
