+++ common/behaviours/CheckListItemBehaviour.php
@@ -3,18 +3,17 @@
 
 namespace common\behaviours;
 
-use common\classes\ConsoleLog;
 use common\models\CheckListItem;
 use common\models\UserInfo;
-use common\RabbitMqHelper\RabbitMqHelper;
+use common\RabbitMqService\RabbitMqService;
 use PhpAmqpLib\Message\AMQPMessage;
 use yii\base\Behavior;
-use PhpAmqpLib\Connection\AMQPStreamConnection;
 use yii\helpers\Json;
 
 
 class CheckListItemBehaviour extends Behavior
 {
+    const QUEUE_FOR_UPDATE_ITEM = 'item_update_queue';
 
     public function events()
     {
@@ -27,21 +26,16 @@ class CheckListItemBehaviour extends Behavior
     {
         /** @var CheckListItem $item */
         $item = $this->owner;
-
-        $channel = RabbitMqHelper::connect('item_update_queue');
-        $msg = new AMQPMessage(
+        /** @var RabbitMqService $rb */
+        $rb = \Yii::$app->rabbitMqService;
+        $rb->sendMessageToQueue(self::QUEUE_FOR_UPDATE_ITEM,
             Json::encode(
                 [
                     "item" => $item->attributes,
                     "dirty" => $item->getDirtyAttributes(),
                     "old" => $item->oldAttributes,
-                ]),
-            [
-                "delivery_mode" => AMQPMessage::DELIVERY_MODE_PERSISTENT
-            ]
+                ]
+            )
         );
-        $channel->publish($msg);
-
-
     }
 }
